requirements.txt
django
celery[redis]
mysqlclient
python-dotenv
# Django REST framework
djangorestframework
markdown
django-filter
